+++ igor-plugin-core/src/main/java/com/arassec/igor/plugin/core/web/connector/HttpConnector.java
@@ -1,11 +1,13 @@
 package com.arassec.igor.plugin.core.web.connector;
 
+import com.arassec.igor.core.model.connector.Connector;
+
 import java.net.http.HttpClient;
 
 /**
  * Interface for HTTP related connectors.
  */
-public interface HttpConnector {
+public interface HttpConnector extends Connector {
 
     /**
      * Returns the configured HTTP client for web requests.